requirements.txt
aiofiles
aiohttp
async-lru==1.0.2
async-timeout==3.0.1
attrs==20.1.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
dnspython==1.16.0
google-api-core==1.22.3
google-api-python-client==1.12.3
google-auth==1.22.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
humanfriendly==8.2
idna==2.10
multidict==4.7.6
oauth2client==4.1.3
oauthlib==3.1.0
pip-autoremove==0.9.1
protobuf==3.13.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.20.0
pymongo==3.11.0
pyreadline==2.1
PySocks==1.7.1
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
TgCrypto~=1.2.3
uritemplate==3.0.1
urllib3==1.25.10
yarl==1.5.1
Pyrogram~=2.0.30
motor
